Move traits into own module and implement Clone for Strategy
