+++ aquadoggo/src/replication/mod.rs
@@ -7,11 +7,12 @@ mod manager;
 mod message;
 mod session;
 mod strategies;
+pub mod traits;
 
 pub use manager::SyncManager;
 pub use message::{StrategyMessage, SyncMessage};
 pub use session::{Session, SessionId, SessionState};
-pub use strategies::{NaiveStrategy, SetReconciliationStrategy, Strategy};
+pub use strategies::{NaiveStrategy, SetReconciliationStrategy, StrategyResult};
 
 /// De-duplicated and sorted set of schema ids which define the target data for the replication
 /// session.
